requirements.txt
# requirements.txt
openai>=1.20.0,<2.0.0
python-dotenv>=0.19.0 
rich>=10.0.0
scikit-learn
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
web3==6.0.0
scipy>=1.7.0
langchain>=0.1.20,<0.2.0
langchain-openai>=0.1.7,<0.2.0
langgraph>=0.0.48,<0.1.0
httpx>=0.27.0
# Pin pydantic versions if needed, these often cause issues
pydantic>=2.7,<2.8
typing_inspect==0.9.0
typing_extensions==4.10.0
simpy==4.0.2

# Add other existing requirements: argparse, etc.

    #   pip install httpx==0.27.2
